# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-garage Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://backend.egarage.store",
            "description": "Dev server online"
        },
        {
            "url": "http://localhost",
            "description": "Dev server offline"
        },
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/v1/cars/add": {
            "post": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Add car.",
                "parameters": [
                    {
                        "parameter": "user_credentials_in_query_required",
                        "name": "body",
                        "in": "query",
                        "description": "Acceptable extensions for thumbnail: png, jpg, jpeg.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Car"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cars/status/{car_id}": {
            "post": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Change availability status of car's.",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarStatusUpdate"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cars": {
            "get": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Get all cars that logged user own.",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cars/update/{car_id}": {
            "put": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Update car's info or thumbnail.",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "user_credentials_in_query_required",
                        "name": "body",
                        "in": "query",
                        "description": "Acceptable extensions for thumbnail: png, jpg, jpeg.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/CarUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cars/update/details/{car_id}": {
            "put": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Update car's details.",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "user_credentials_in_query_required",
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/CarDetailsUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cars/delete/{car_id}": {
            "delete": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Delete car.",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "properties": {
                    "brand": {
                        "type": "string|required"
                    },
                    "description": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "file"
                    }
                },
                "type": "object",
                "example": {
                    "brand": "BMW X12",
                    "description": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam tempora aperiam sint sequi.",
                    "thumbnail": "file"
                }
            },
            "CarCollection": {
                "properties": {
                    "cars": {
                        "type": "list"
                    }
                },
                "type": "object",
                "example": [
                    {
                        "brand": "BMW X12",
                        "description": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam tempora aperiam sint sequi.",
                        "thumbnail": "file"
                    },
                    {
                        "brand": "BMW X12",
                        "description": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam tempora aperiam sint sequi.",
                        "thumbnail": "file"
                    }
                ]
            },
            "CarUpdate": {
                "properties": {
                    "brand": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "file"
                    }
                },
                "type": "object",
                "example": {
                    "brand": "BMW X12",
                    "description": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam tempora aperiam sint sequi.",
                    "thumbnail": "file"
                }
            },
            "CarDetailsUpdate": {
                "properties": {
                    "engine_capacity": {
                        "type": "string"
                    },
                    "horsePower": {
                        "type": "string"
                    },
                    "sits": {
                        "type": "string"
                    },
                    "doors": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "drivetrain": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "Fuel_Type": {
                        "type": "string"
                    },
                    "mileage": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "engine_capacity": "2.0l",
                    "horse_power": "200hp",
                    "sits": "5",
                    "doors": "5",
                    "color": "silver",
                    "drivetrain": "FWD",
                    "body": "suv",
                    "Feul_Type": "petrol",
                    "mileage": "150000km"
                }
            },
            "CarStatusUpdate": {
                "properties": {
                    "brand": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "file"
                    },
                    "availability": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "brand": "BMW X12",
                    "description": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam tempora aperiam sint sequi.",
                    "thumbnail": "file",
                    "availability": "available"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}